Time: 19 ms (9.2%), Space: 42.6 MB (18.57%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,9 +1,9 @@
 class Solution {
     public int mySqrt(int x) {
-        int res = 0;
+        long res = 0;
         while(res*res <= x) 
             res++;
 
-        return res-1;
+        return (int)res-1;
     }
 }
